LPS-181172 Move classes to manager package
